+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 - Skip health check access logs.
 
+- Fix access logging of correlation id.
+
 
 0.8.1 (2023-11-06)
 ------------------